added a way to determin if a item should be displayed on the slide show

+++ models/model.rb
@@ -3,5 +3,5 @@ class Item
 
   property :id, Serial
   property :name, String
-  property :price, Decimal
+  property :price, Float
 end
